requirements.txt
streamlit==1.35.0
openai==0.28.1
python-dotenv==1.0.1
googletrans==4.0.0-rc1
